Tests/LibWeb/Text/expected/css/supports-other-at-rules.txt

@supports at-rule(@container): true
@supports at-rule(@font-face): true
@supports at-rule(@font-feature-values): true
@supports at-rule(@function): true
@supports at-rule(@keyframes): true
@supports at-rule(@-webkit-keyframes): true
@supports at-rule(@layer): true
@supports at-rule(@namespace): true
@supports at-rule(@page): true
@supports at-rule(@property): true
@supports at-rule(@scope): true
@supports at-rule(@stylistic): true
@supports at-rule(@historical-forms): true
@supports at-rule(@styleset): true
@supports at-rule(@character-variant): true
@supports at-rule(@swash): true
@supports at-rule(@ornaments): true
@supports at-rule(@annotation): true
@supports at-rule(@top-left-corner): true
@supports at-rule(@top-left): true
@supports at-rule(@top-center): true
@supports at-rule(@top-right): true
@supports at-rule(@top-right-corner): true
@supports at-rule(@bottom-left-corner): true
@supports at-rule(@bottom-left): true
@supports at-rule(@bottom-center): true
@supports at-rule(@bottom-right): true
@supports at-rule(@bottom-right-corner): true
@supports at-rule(@left-top): true
@supports at-rule(@left-middle): true
@supports at-rule(@left-bottom): true
@supports at-rule(@right-top): true
@supports at-rule(@right-middle): true
@supports at-rule(@right-bottom): true
